+++ frontend/src/lib/components/Sharing/SharingModal.tsx
@@ -9,7 +9,7 @@ import { copyToClipboard } from 'lib/utils'
 import { IconGlobeLock, IconInfo, IconLink, IconLockLemon, IconUnfoldLess, IconUnfoldMore } from '../icons'
 import { CodeSnippet, Language } from 'scenes/ingestion/frameworks/CodeSnippet'
 import { DashboardCollaboration } from 'scenes/dashboard/DashboardCollaborators'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { Tooltip } from 'lib/components/Tooltip'
 import './SharingModal.scss'
 import { Form } from 'kea-forms'
