+++ frontend/src/lib/components/Subscriptions/views/EditSubscription.tsx
@@ -3,7 +3,7 @@ import { useActions, useValues } from 'kea'
 import { LemonButton } from 'lib/components/LemonButton'
 import { membersLogic } from 'scenes/organization/Settings/membersLogic'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { dayjs } from 'lib/dayjs'
 import { LemonSelect, LemonSelectOptions, LemonSelectProps } from 'lib/components/LemonSelect'
 import { subscriptionLogic } from '../subscriptionLogic'
