+++ frontend/src/lib/forms/Field.tsx
@@ -1,81 +1,46 @@
-import { FieldProps as KeaFieldProps } from 'kea-forms'
-import clsx from 'clsx'
-import { Field as KeaField } from 'kea-forms/lib/components'
+import { IconErrorOutline } from 'lib/components/icons'
 import React from 'react'
+import { LemonLabel } from '../components/LemonLabel/LemonLabel'
+import { Field as KeaField, FieldProps as KeaFieldProps } from 'kea-forms/lib/components'
 
-export interface FieldProps extends KeaFieldProps {
-    showOptional?: boolean
-    className?: string
-    style?: React.CSSProperties
-    help?: string
+export type PureFieldProps = {
+    /** The label name to be displayed */
+    label?: React.ReactNode
+    /** Info tooltip to be displayed next to the label */
+    info?: React.ReactNode
+    /** Help text to be shown directly beneath the input */
+    help?: React.ReactNode
+    /** Error message to be displayed */
+    error?: React.ReactNode
+    children?: React.ReactNode
 }
 
-/** Compatible replacement for Form.Item. Passes `value` and `onChange(value: any)` to its children. */
-export const Field = ({
-    showOptional,
-    name,
-    help,
-    className,
-    style,
-    noStyle,
-    ...keaFieldProps
-}: FieldProps): JSX.Element => {
-    /** Drop-in replacement antd template for kea forms */
-    const template: FieldProps['template'] = noStyle
-        ? ({ kids }) => <>{kids}</>
-        : ({ label, kids, hint, error }) => {
-              return (
-                  <div
-                      className={clsx(
-                          'ant-row',
-                          'ant-form-item',
-                          help || hint || error ? 'ant-form-item-with-help' : '',
-                          error ? `ant-form-item-has-error` : '',
-                          className
-                      )}
-                      style={style}
-                  >
-                      {label ? (
-                          <div className="ant-col ant-form-item-label">
-                              <label htmlFor={String(name)} title={typeof label === 'string' ? label : undefined}>
-                                  {showOptional ? (
-                                      <>
-                                          {label}
-                                          {showOptional ? (
-                                              <span className="ant-form-item-optional" title="">
-                                                  (optional)
-                                              </span>
-                                          ) : null}
-                                      </>
-                                  ) : (
-                                      label
-                                  )}
-                              </label>
-                          </div>
-                      ) : null}
-                      <div className="ant-col ant-form-item-control">
-                          <div className="ant-form-item-control-input">
-                              <div className="ant-form-item-control-input-content">{kids}</div>
-                          </div>
-                          {hint || error || help ? (
-                              <div className="ant-form-item-explain ant-form-item-explain-connected">
-                                  {error ? (
-                                      <div role="alert" className="ant-form-item-explain-error">
-                                          {error}
-                                      </div>
-                                  ) : null}
-                                  {hint ? (
-                                      <div role="alert" className="ant-form-item-explain-warning">
-                                          {hint}
-                                      </div>
-                                  ) : null}
-                                  {help ? <div className="ant-form-item-explain">{help}</div> : null}
-                              </div>
-                          ) : null}
-                      </div>
-                  </div>
-              )
-          }
+/** A "Pure" field - used when you want the Field styles without the Kea form functionality */
+export const PureField = ({ label, info, error, help, children }: PureFieldProps): JSX.Element => {
+    return (
+        <div className="flex flex-col gap-2">
+            {label ? <LemonLabel info={info}>{label}</LemonLabel> : null}
+            {children}
+            {help ? <div className="text-muted text-xs">{help}</div> : null}
+            {error ? (
+                <div className="text-danger flex items-center gap-1 text-sm">
+                    <IconErrorOutline className="text-xl" /> {error}
+                </div>
+            ) : null}
+        </div>
+    )
+}
 
-    return <KeaField {...keaFieldProps} name={name} template={template} />
+export type FieldProps = Omit<PureFieldProps, 'children' | 'error'> & Pick<KeaFieldProps, 'children' | 'name'>
+
+export const Field = ({ name, help, ...keaFieldProps }: FieldProps): JSX.Element => {
+    /** Drop-in replacement antd template for kea forms */
+    const template: KeaFieldProps['template'] = ({ label, kids, error }) => {
+        return (
+            <PureField label={label} error={error} help={help}>
+                {kids}
+            </PureField>
+        )
+    }
+    return <KeaField {...keaFieldProps} name={name} template={template} noStyle />
 }
