+++ frontend/src/scenes/actions/ActionEdit.tsx
@@ -15,7 +15,7 @@ import { EditableField } from 'lib/components/EditableField/EditableField'
 import { ActionStepType, AvailableFeature } from '~/types'
 import { userLogic } from 'scenes/userLogic'
 import { ObjectTags } from 'lib/components/ObjectTags/ObjectTags'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { LemonButton } from 'lib/components/LemonButton'
 import { LemonCheckbox } from 'lib/components/LemonCheckbox'
 import { LemonInput } from 'lib/components/LemonInput/LemonInput'
