+++ frontend/src/scenes/cohorts/CohortEdit.tsx
@@ -7,7 +7,7 @@ import { LemonButton } from 'lib/components/LemonButton'
 import { router } from 'kea-router'
 import { urls } from 'scenes/urls'
 import { Divider } from 'antd'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { LemonInput } from 'lib/components/LemonInput/LemonInput'
 import { Tooltip } from 'lib/components/Tooltip'
 import { LemonSelect } from 'lib/components/LemonSelect'
