+++ frontend/src/scenes/data-management/definition/DefinitionEdit.tsx
@@ -5,7 +5,7 @@ import { useActions, useValues } from 'kea'
 import { definitionEditLogic, DefinitionEditLogicProps } from 'scenes/data-management/definition/definitionEditLogic'
 import { LemonButton } from 'lib/components/LemonButton'
 import { Col, Divider, Row } from 'antd'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { LemonInput } from 'lib/components/LemonInput/LemonInput'
 import { LemonTextArea } from 'lib/components/LemonTextArea/LemonTextArea'
 import { ObjectTags } from 'lib/components/ObjectTags/ObjectTags'
