+++ frontend/src/scenes/instance/AsyncMigrations/AsyncMigrationParametersModal.tsx
@@ -4,9 +4,8 @@ import { AsyncMigrationModalProps, asyncMigrationsLogic } from 'scenes/instance/
 import { LemonModal } from 'lib/components/LemonModal/LemonModal'
 import { LemonButton } from 'lib/components/LemonButton'
 import { asyncMigrationParameterFormLogic } from 'scenes/instance/AsyncMigrations/asyncMigrationParameterFormLogic'
-import { Field } from 'kea-forms'
+import { Field, Form } from 'kea-forms'
 import { LemonInput } from 'lib/components/LemonInput/LemonInput'
-import { VerticalForm } from 'lib/forms/VerticalForm'
 import { AnimatedCollapsible } from 'lib/components/AnimatedCollapsible'
 
 export function AsyncMigrationParametersModal(props: AsyncMigrationModalProps): JSX.Element {
@@ -15,7 +14,7 @@ export function AsyncMigrationParametersModal(props: AsyncMigrationModalProps): 
     const [collapsed, setCollapsed] = useState(true)
 
     return (
-        <VerticalForm
+        <Form
             logic={asyncMigrationParameterFormLogic}
             props={props}
             formKey="parameters"
@@ -73,6 +72,6 @@ export function AsyncMigrationParametersModal(props: AsyncMigrationModalProps): 
                     ))}
                 </AnimatedCollapsible>
             </LemonModal>
-        </VerticalForm>
+        </Form>
     )
 }
