+++ frontend/src/scenes/organization/ConfirmOrganization/ConfirmOrganization.tsx
@@ -8,7 +8,7 @@ import { WelcomeLogo } from 'scenes/authentication/WelcomeLogo'
 import { LemonDivider } from 'lib/components/LemonDivider'
 import { useActions, useValues } from 'kea'
 import { confirmOrganizationLogic } from './confirmOrganizationLogic'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { AnimatedCollapsible } from 'lib/components/AnimatedCollapsible'
 import { urls } from 'scenes/urls'
 import { Form } from 'kea-forms'
