+++ frontend/src/scenes/organization/Settings/VerifiedDomains/ConfigureSAMLModal.tsx
@@ -4,7 +4,7 @@ import { LemonModal } from 'lib/components/LemonModal/LemonModal'
 import React from 'react'
 import { verifiedDomainsLogic } from './verifiedDomainsLogic'
 import { Form } from 'kea-forms'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { AlertMessage } from 'lib/components/AlertMessage'
 import { LemonInput } from 'lib/components/LemonInput/LemonInput'
 import { LemonTextArea } from 'lib/components/LemonTextArea/LemonTextArea'
