+++ frontend/src/scenes/plugins/source/PluginSource.tsx
@@ -8,7 +8,7 @@ import { Drawer } from 'lib/components/Drawer'
 import { userLogic } from 'scenes/userLogic'
 import { canGloballyManagePlugins } from '../access'
 import { pluginSourceLogic } from 'scenes/plugins/source/pluginSourceLogic'
-import { Field } from 'lib/forms/FieldV2'
+import { Field } from 'lib/forms/Field'
 import { PluginSourceTabs } from 'scenes/plugins/source/PluginSourceTabs'
 import { LemonButton } from 'lib/components/LemonButton'
 import { createDefaultPluginSource } from 'scenes/plugins/source/createDefaultPluginSource'
